6.3. Массив заполнен true или false [#53857][#91116]

+++ chapter_001/src/main/java/ru/job4j/array/Turn.java
@@ -14,7 +14,7 @@ package ru.job4j.array;
 *@return array - начальный массив с переставленными в обратном порядке элементами.
 */		
     public int[] back(int[] array) {
-        for (int i = 0; i < (int) (array.length / 2); i++) {
+        for (int i = 0; i < array.length / 2; i++) {
 			int temp = array[array.length - i - 1];
 			array[array.length - i - 1] = array[i];
 			array[i] = temp;
